Use labels for choice questions and add helpers
For whatever reason, the label field was excluded from the admin forms
for single- & multiple-choice questions, but they have now been
reinstated.

Static helper methods have been added to the Question class in order to
help find all questions on a given page or in a hierarchy of pages. A
new helper was also added to help provide scoring for a query set of
questions (to be used in conjunction with the other new helpers).